+++ src/function/methods.js
@@ -1,6 +1,8 @@
 export { default as debounce } from './debounce';
+export { default as lock } from './lock';
 export { default as memoize } from './memoize';
 export { default as once } from './once';
+export { default as partial } from './partial';
 export { default as setInterval } from './setInterval';
 export { default as setTimeout } from './setTimeout';
 export { default as throttle } from './throttle';